refactor(OutputView): 결과 생성 구현 분리
결과 생성 구현 분리
- 기존에 OutputView에 존재했던 결과 생성 책임을 RaceManager에게 위임
