+++ src/main/java/racingcar/view/OutputView.java
@@ -1,8 +1,5 @@
 package racingcar.view;
 
-import java.util.List;
-import racingcar.model.Car;
-
 public class OutputView {
 
     private static final String ROUND_RESULT_HEADER = "실행 결과";
@@ -12,31 +9,11 @@ public class OutputView {
         System.out.println(ROUND_RESULT_HEADER);
     }
 
-    public void printRoundResult(List<Car> cars) {
-        StringBuilder stringBuilder = new StringBuilder();
-
-        cars.forEach(car -> {
-            stringBuilder.append(car.getCarName());
-            stringBuilder.append(" : ");
-            stringBuilder.append("-".repeat(Math.max(0, car.getCurrentLocation())));
-            stringBuilder.append("\n");
-        });
-
-        System.out.println(stringBuilder);
+    public void printRoundResult(String roundResult) {
+        System.out.println(roundResult);
     }
 
-    public void printWinner(List<Car> winners) {
-        StringBuilder stringBuilder = new StringBuilder();
-
-        stringBuilder.append(WINNER_HEADER);
-
-        for (int i = 0; i < winners.size(); i++) {
-            stringBuilder.append(winners.get(i).getCarName());
-            if (i < winners.size() - 1) {
-                stringBuilder.append(", ");
-            }
-        }
-
-        System.out.println(stringBuilder);
+    public void printWinners(String winners) {
+        System.out.println(WINNER_HEADER + winners);
     }
 }
